Monte Carlo Integration

Simple Monte Carlo Integraton

mcint(f, a, b, m = 1000)

mcint2(f, xdom, ydom, m = 1000)

Arguments

f
function to integrate

a
the lower-bound of integration

b
the upper-bound of integration

m
the number of subintervals to calculate

xdom
the domain on x of integration in two dimensions

ydom
the domain on y of integration in two dimensions

Value

the value of the integral

Details

The mcint function uses a simple Monte Carlo algorithm to estimate the value of an integral. The parameter n sets the total number of evaluation points. The parameter max.y is the maximum expected value of the range of function f. The mcint2 provides Monte Carlo integration in two dimensions.

See also

Examples

f <- function(x) { sin(x)^2 + log(x)}
mcint(f, 0, 1)
#> [1] -0.7141941
mcint(f, 0, 1, m = 10e6)
#> [1] -0.7267529