Wave Equation using

solve heat equation via forward-time central-space method

wave(u, alpha, xdelta, tdelta, n)

Arguments

u
the initial values of u

alpha
the thermal diffusivity coefficient

xdelta
the change in x at each step in u

tdelta
the time step

n
the number of steps to take

Value

a matrix of u values at each time step

Details

The heat solves the heat equation using the forward-time central-space method in one-dimension.

Examples

speed <- 2
x0 <- 0
xdelta <- .05
x <- seq(x0, 1, xdelta)
m <- length(x)
u <- sin(x * pi * 2)
u[11:21] <- 0
tdelta <- .02
n <- 40
z <- wave(u, speed, xdelta, tdelta, n)